requirements.txt
# Core Flask & Authentication
flask==2.3.2
flask-jwt-extended==4.5.2
python-dotenv==1.0.0

# Database & Caching
redis==4.6.0
sqlalchemy==2.0.20  # Optional (if adding PostgreSQL later)

# Machine Learning
scikit-learn==1.3.0
joblib==1.3.2
pandas==2.1.1
numpy==1.26.0

# OAuth & Security
authlib==1.2.1
requests==2.31.0

# Production Deployment
gunicorn==21.2.0